simple_baseline/requirements.txt
numpy
opencv-python
scipy
tensorflow-gpu
tqdm
